moved helper functions to loadprompt_helper for clearer file
