+++ cmd/holoai/main.go
@@ -9,7 +9,6 @@ import (
 	"github.com/METIL-HoloAI/HoloTable-Middleware/internal/config"
 	"github.com/METIL-HoloAI/HoloTable-Middleware/internal/database"
 	"github.com/METIL-HoloAI/HoloTable-Middleware/internal/listeners"
-
 	_ "github.com/mattn/go-sqlite3"
 )
